Remove duplicated code by creating a super class for the metrics
