+++ pywatts/modules/rolling_mae.py
@@ -1,15 +1,10 @@
-from typing import Dict
-
 import numpy as np
 import pandas as pd
-import xarray as xr
 
-from pywatts.core.base import BaseTransformer
-from pywatts.core.exceptions.input_not_available import InputNotAvailable
-from pywatts.utils._xarray_time_series_utils import _get_time_indeces
+from pywatts.modules.rolling_metric_base import RollingMetricBase
 
 
-class RollingMAE(BaseTransformer):
+class RollingMAE(RollingMetricBase):
     """
     Module to calculate the Rolling Mean Absolute Error (MAE)
     :param window_size: Determine the window size for the rolling mae. Default 24
@@ -19,67 +14,8 @@ class RollingMAE(BaseTransformer):
 
     """
 
-    def __init__(self, name: str = "RollingMAE", window_size=24, window_size_unit="d"):
-        super().__init__(name)
-        self.window_size_unit = window_size_unit
-        self.window_size = window_size
-
-    def get_params(self) -> Dict[str, object]:
-        """
-        Returns a dict of parameters used in the RollingMAE.
-
-        :return: Parameters set for the RollingMAE
-        :rtype: Dict[str, object]
-        """
-        return {
-            "window_size_unit": self.window_size_unit,
-            "window_size": self.window_size
-        }
-
-    def set_params(self, window_size=None, window_size_unit=None):
-        """
-        Set the parameter for the RollingMAE.
-
-        :param window_size: Determine the window size if a rolling mae should be calculated.
-                            Ignored if rolling is set to False. Default 24
-        :type window_size: int
-        :param window_size_unit: Determine the unit of the window size. Default Day (d)"
-        :type window_size_unit: str
-
-        """
-        if window_size is not None:
-            self.window_size = window_size
-        if window_size_unit is not None:
-            self.window_size_unit = window_size_unit
-
-    def transform(self, y: xr.DataArray, **kwargs: xr.DataArray) -> xr.DataArray:
-        """
-        Calculates the MAE based on the predefined target and predictions variables.
-
-        :param x: the input dataset
-        :type x: Optional[xr.DataArray]
-
-        :return: The calculated MAE
-        :rtype: xr.DataArray
-        """
-
-        if kwargs == {}:
-            error_message = "No predictions are provided as input for the RollingMAE. You should add the predictions" \
-                            " by a seperate key word arguments if you add the RollingMAE to the pipeline."
-            self.logger.error(error_message)
-            raise InputNotAvailable(error_message)
-
-        t = y.values
-        results = {}
-        for key, y_hat in kwargs.items():
-            p = y_hat.values
-            p_, t_ = p.reshape((len(p), -1)), t.reshape((len(t), -1))
-            results[key] = pd.DataFrame(np.mean((p_ - t_), axis=-1),
-                                        index=y.indexes[_get_time_indeces(kwargs)[0]]).rolling(
-                f"{self.window_size}{self.window_size_unit}").apply(
-                lambda x: np.mean(np.abs(x))).values
-        time = y.indexes[_get_time_indeces(y)[0]]
-
-        return xr.DataArray(np.concatenate(list(results.values()), axis=1),
-                            coords={"time": time, "predictions": list(results.keys())},
-                            dims=["time", "predictions"])
+    def _apply_rolling_metric(self, p, t, index):
+        return pd.DataFrame(np.mean((p - t), axis=-1),
+                                    index=index).rolling(
+            f"{self.window_size}{self.window_size_unit}").apply(
+            lambda x: np.mean(np.abs(x))).values
