+++ tests/unit/summaries/test_mean_absolute_error.py
@@ -85,7 +85,7 @@ class TestMAE(unittest.TestCase):
                          "seperate key word arguments if you add the MAE to the pipeline.")
 
     @patch("builtins.open")
-    @patch("pywatts.summaries.mae_summary.cloudpickle")
+    @patch("pywatts.summaries.metric_base.cloudpickle")
     def test_save(self, cloudpickle_mock, open_mock):
         fm_mock = MagicMock()
         fm_mock.get_path.return_value = "filter_path"
@@ -104,7 +104,7 @@ class TestMAE(unittest.TestCase):
 
 
     @patch("builtins.open")
-    @patch("pywatts.summaries.mae_summary.cloudpickle")
+    @patch("pywatts.summaries.metric_base.cloudpickle")
     def test_load(self, cloudpickle_mock, open_mock):
         load_information = {'params': {'offset': 24}, 'name': 'NAME', 'class': 'RMSE',
                             'module': 'pywatts.summaries.mae_summary', 'filter': 'filter_path'}
